removing many rules. Filter script.
Some of these sites may be dead. They may also still be around and I
just have not seen them today. Yay to git commit history!

To filter down the ruleset, turn on uBlockss logs and navigate to a
bunch of sites.  Then filter those logs down to the unique rules we've
seen.  This then is the found.txt file, which feeds into filter.sh,
which removes any rules we did not use.